+++ CHANGES.md
@@ -23,6 +23,7 @@ To be released.
 
  -  Added `BlockMetadata` class.  [[#1164], [#1492]]
  -  Added `BlockContent<T>` class.  [[#1164], [#1492]]
+ -  Added `PreEvaluationBlockHeader` class.  [[#1146], [#1164], [#1492]]
  -  Added `PreEvaluationBlock<T>` class.  [[#1146], [#1164], [#1492]]
  -  Added `BlockDigest.FromBlock<T>()` static method. [[#1492]]